+++ NEWS.md
@@ -1,5 +1,8 @@
 # cols4all 0.7
 - added naming column
+- improved similarity matrix plot
+- added new 'tol' palettes
+- scores shown with more precision
 
 # cols4all 0.6-1
 - fixed bug confusion lines